+++ src/lib/index.ts
@@ -1,7 +1,15 @@
-export { default as BaseTetherTooltip } from './BaseTetherTooltip.svelte';
-export type { BaseTetherTooltipState } from './BaseTetherTooltip.svelte';
-export { default as Portal } from './Portal.svelte';
-export { default as PortalOverlay } from './PortalOverlay.svelte';
-export { default as Tether } from './Tether.svelte';
-export type { Alignment, SizeInheritMode, TetherState } from './Tether.svelte';
-export { default as TetherBoundary } from './TetherBoundary.svelte';
+export { default as BaseTetherTooltip } from './patterns/BaseTetherTooltip.svelte';
+export type {
+	BaseTetherTooltipFocusState,
+	BaseTetherTooltipState
+} from './patterns/BaseTetherTooltip.svelte';
+export { mountSnippet } from './portaling/mount-snippet.ts';
+export { default as Portal } from './portaling/Portal.svelte';
+export { default as PortalOverlay } from './portaling/PortalOverlay.svelte';
+export type { SizeInheritMode, TetherState } from './tethering/common-types.ts';
+export { tether } from './tethering/tether-attachment.svelte.ts';
+export type { TetherAttachmentOptions } from './tethering/tether-attachment.svelte.ts';
+export type { Alignment } from './tethering/tether-layout.ts';
+export { default as Tether } from './tethering/Tether.svelte';
+export type { TetherProps } from './tethering/Tether.svelte';
+export { default as TetherBoundary } from './tethering/TetherBoundary.svelte';